+++ guniflask/config/__init__.py
@@ -5,6 +5,5 @@ from .app_settings import settings
 from .env import app_name_from_env
 from .env import load_app_env
 from .env import set_app_default_env
-from .loader import load_app_settings
 from .loader import load_config
 from .loader import load_profile_config