+++ guniflask/security_config/__init__.py
@@ -1,7 +1,5 @@
 # coding=utf-8
 
-from .annotation import enable_authentication
-from .annotation import enable_web_security
 from .authentication_manager_builder import AuthenticationManagerBuilder
 from .dao_authentication_configurer import DaoAuthenticationConfigurer
 from .http_security import HttpSecurity