+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Changelog
 - BaseModelMixin声明query的类型
 - 提供接口级别单元测试方案
 - 移除oauth2相关功能，后续身份认证、授权等相关功能将基于Keycloak实现
+- ``current_user`` 的定义移动到 ``guniflask.security`` 模块中
 
 0.10.0 (2020-10-19)
 ------------------